Renamed Encoder to QuadratureEncoder
Resolves #2185
Moved Encoder functionality to QuadratureEncoder.java, leaving Encoder.java as a deprecated wrapper class.
